break out top row and start cleanup
