+++ packages/commonwealth/client/styles/pages/user_dashboard/dashboard_communities_preview.scss
@@ -22,35 +22,18 @@
 }
 
 .DashboardCommunitiesPreview {
+  align-items: center;
   display: flex;
   flex-direction: column;
-  gap: 8px;
+  gap: 16px;
 
   @include smallInclusive {
     display: none;
   }
 
   .community-preview-cards-collection {
-    align-items: center;
     display: flex;
     flex-direction: column;
     gap: 16px;
-
-    .view-more-communities-link {
-      align-items: center;
-      background-color: $purple-50;
-      border-radius: $border-radius-corners;
-      color: $purple-400;
-      cursor: pointer;
-      display: flex;
-      height: 40px;
-      padding: 4px 16px;
-      width: fit-content;
-
-      &:hover {
-        background-color: $purple-100;
-        text-decoration: none;
-      }
-    }
   }
 }
