+++ packages/commonwealth/client/styles/pages/user_dashboard/index.scss
@@ -15,7 +15,7 @@
     .dashboard-header {
       display: flex;
       flex-direction: column;
-      gap: 8px;
+      gap: 16px;
       width: 100%;
     }
   }
