+++ packages/commonwealth/client/styles/pages/user_dashboard/user_dashboard_row_top.scss
@@ -1,0 +1,59 @@
+@import 'shared';
+
+.UserDashboardRowTop {
+  display: flex;
+
+  span.header {
+    font-weight: 600;
+
+    span.community-title {
+      font-weight: 500;
+    }
+
+    span.comment-counts {
+      color: $neutral-600;
+      font-weight: 400;
+    }
+
+    span.community-link {
+      font-weight: 600;
+      cursor: pointer;
+
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+
+    // span.block-number {
+    //   color: $neutral-600;
+    //   font-weight: 400;
+    //   margin-left: 20px;
+    // }
+  }
+
+  //   .comment-body {
+  //     margin: 7px 0 8px 25px;
+
+  //     &-concat {
+  //       text-overflow: ellipsis;
+  //       overflow: hidden;
+  //       display: -webkit-box;
+  //       -webkit-line-clamp: 3;
+  //       -webkit-box-orient: vertical;
+  //       margin: 7px 0 8px 25px;
+
+  //       .MarkdownFormattedText {
+  //         white-space: inherit !important;
+  //       }
+  //     }
+  //   }
+
+  //   .event-body {
+  //     text-overflow: ellipsis;
+  //     overflow: hidden;
+  //     display: -webkit-box;
+  //     -webkit-line-clamp: 2;
+  //     -webkit-box-orient: vertical;
+  //     margin: 7px 0 8px 0px;
+  //   }
+}
